These changes account moving the makeThings and GoogleCalendar modules in their own directories and the corresponding changes needed the import statements.
