+++ StateController.py
@@ -1,6 +1,5 @@
-# from datetime import datetime
-import things
 import logging
+import things
 
 
 class State:
@@ -28,44 +27,44 @@ class State:
 			return True
 
 
-	def list_updated_deadlines(self, updated_deadlines: list[dict]) -> list[dict]:
-		updated_deadlines_list: list[dict] = []
+	# def list_updated_deadlines(self, updated_deadlines: list[dict]) -> list[dict]:
+	# 	updated_deadlines_list: list[dict] = []
 
-		for deadline in updated_deadlines:
-			# First check to see if the task is in current_tasks already. 
-			state_deadline = [i for i in self.current_deadlines if i['uuid'] == deadline['uuid']]
-			if not state_deadline:
-				# If a task is totally new, then it's ID won't be in self.current_tasks yet and we can skip it.
-				pass 
-			elif deadline == state_deadline[0]:
-				# If the state_task and updated task are the same then we don't need to update anything
-				pass
-			elif not state_deadline[0].get('deadline'):
-				# No need to update the calendar if there is no deadline. 
-				pass
-			elif not deadline.get('deadline'):
-				# If the deadline has been deleted from the task then we skip it here. 
-				pass
-			elif deadline.get('status') == 'completed':
-				# If the task is already completed we can pass on updating it. 
-				pass
+	# 	for deadline in updated_deadlines:
+	# 		# First check to see if the task is in current_tasks already. 
+	# 		state_deadline = [i for i in self.current_deadlines if i['uuid'] == deadline['uuid']]
+	# 		if not state_deadline:
+	# 			# If a task is totally new, then it's ID won't be in self.current_tasks yet and we can skip it.
+	# 			pass 
+	# 		elif deadline == state_deadline[0]:
+	# 			# If the state_task and updated task are the same then we don't need to update anything
+	# 			pass
+	# 		elif not state_deadline[0].get('deadline'):
+	# 			# No need to update the calendar if there is no deadline. 
+	# 			pass
+	# 		elif not deadline.get('deadline'):
+	# 			# If the deadline has been deleted from the task then we skip it here. 
+	# 			pass
+	# 		elif deadline.get('status') == 'completed':
+	# 			# If the task is already completed we can pass on updating it. 
+	# 			pass
 
-			else:
-				# Only update this event if one of these fields specifically has changed 
-				state_deadline = state_deadline[0]
-				state_task_values = {'title': state_deadline.get('title'),
-									 'uuid': state_deadline.get('uuid'),
-									 'deadline': state_deadline.get('deadline')}
+	# 		else:
+	# 			# Only update this event if one of these fields specifically has changed 
+	# 			state_deadline = state_deadline[0]
+	# 			state_task_values = {'title': state_deadline.get('title'),
+	# 								 'uuid': state_deadline.get('uuid'),
+	# 								 'deadline': state_deadline.get('deadline')}
 				
-				updated_task_values = {'title': deadline.get('title'),
-									   'uuid': deadline.get('uuid'),
-									   'deadline': deadline.get('deadline')}
+	# 			updated_task_values = {'title': deadline.get('title'),
+	# 								   'uuid': deadline.get('uuid'),
+	# 								   'deadline': deadline.get('deadline')}
 				
-				if state_task_values != updated_task_values:
-					logging.debug(f"UPDATED DEADLINE FOUND: {deadline.get('uuid')} | {deadline.get('title')}")
-					deadline.update({'change_type': 'update_deadline'})
-					updated_deadlines_list.append(deadline)
-				else:
-					logging.debug("No updated deadlines found")
+	# 			if state_task_values != updated_task_values:
+	# 				logging.debug(f"UPDATED DEADLINE FOUND: {deadline.get('uuid')} | {deadline.get('title')}")
+	# 				deadline.update({'change_type': 'update_deadline'})
+	# 				updated_deadlines_list.append(deadline)
+	# 			else:
+	# 				logging.debug("No updated deadlines found")
 
-		return updated_deadlines_list
+	# 	return updated_deadlines_list
